requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.5.2
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
aioboto3==12.3.0
python-dotenv==1.0.0
passpy==0.1.0
qrcode[pil]==7.4.2
pillow==10.1.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
gunicorn==21.2.0 